+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Cahngelog
 
+## 0.1.1
+
+- Fixed error in powerOfTen() method
+- New scale() Method
+- Updated documentation
+
 ## 0.1.0
 
 - Initial version.